pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cageinator"
dynamic = ["version"]
description = "Cageinator: batch pipeline from linkers and nodes to assembled cages."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "TBD" }
authors = [{ name = "Jan Leodolter" }]
keywords = ["chemistry", "computational-chemistry", "xtb", "openbabel", "cage", "cheminformatics", "coordination"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Operating System :: OS Independent",
]

dependencies = [
  "openbabel-wheel>=3.1.1.22",
  "numpy",
]

[project.scripts]
cageinator = "cageinator:main"

[tool.setuptools]
include-package-data = true
py-modules = ["cageinator"]

[tool.setuptools.packages.find]
where = ["."]
include = ["cage_assemblers*", "calculation*", "interfaces*", "utility_functions*"]

[tool.setuptools.dynamic]
version = { attr = "cageinator.__version__" }

[tool.setuptools.package-data]
calculation = ["xtb_opt.sh"]